chore: add internal utility functions `ensureHandlerRef` and `ensureControllerRef`
