+++ packages/core/src/decorators/request-body.decorator.ts
@@ -1,6 +1,6 @@
-import { Auralis, AURALIS_REGISTRY_SYMBOL } from "../auralis.ts";
 import type { Constructor } from "../utilities/constructor.util.ts";
 import { getParamNames } from "../utilities/param-names.util.ts";
+import { ensureHandlerRef } from "../utilities/registry.util.ts";
 
 export function RequestBody(type: Constructor): ParameterDecorator {
   return function (target, propertyKey, parameterIndex) {
@@ -11,18 +11,8 @@ export function RequestBody(type: Constructor): ParameterDecorator {
     const paramNames = getParamNames(fn);
     const paramName: string = paramNames[parameterIndex];
 
-    if (!Auralis[AURALIS_REGISTRY_SYMBOL].has(controller)) {
-      Auralis[AURALIS_REGISTRY_SYMBOL].set(controller, {});
-    }
+    const { handlerRef } = ensureHandlerRef(controller, fn);
 
-    const controllerRef = Auralis[AURALIS_REGISTRY_SYMBOL].get(controller)!;
-    controllerRef.handlers ??= new Map();
-
-    if (!controllerRef.handlers.has(fn)) {
-      controllerRef.handlers.set(fn, {});
-    }
-
-    const handlerRef = controllerRef.handlers.get(fn)!;
     handlerRef.requestBody = {
       paramName,
       type,
@@ -31,9 +21,9 @@ export function RequestBody(type: Constructor): ParameterDecorator {
 
     if (process.env.AURALIS_DEBUG) {
       console.debug("[RequestBody]:", {
-        owningClass: controller,
-        propertyKey,
-        parameterIndex,
+        args: [target, propertyKey, parameterIndex],
+        controller,
+        fn,
         type,
         paramName,
       });
