+++ packages/core/src/decorators/response.decorator.ts
@@ -1,6 +1,6 @@
-import { Auralis, AURALIS_REGISTRY_SYMBOL } from "../auralis.ts";
 import type { Constructor } from "../utilities/constructor.util.ts";
 import { getParamNames } from "../utilities/param-names.util.ts";
+import { ensureHandlerRef } from "../utilities/registry.util.ts";
 
 export const Response: ParameterDecorator = (
   target,
@@ -14,18 +14,8 @@ export const Response: ParameterDecorator = (
   const paramNames = getParamNames(fn);
   const paramName: string = paramNames[parameterIndex];
 
-  if (!Auralis[AURALIS_REGISTRY_SYMBOL].has(controller)) {
-    Auralis[AURALIS_REGISTRY_SYMBOL].set(controller, {});
-  }
+  const { handlerRef } = ensureHandlerRef(controller, fn);
 
-  const controllerRef = Auralis[AURALIS_REGISTRY_SYMBOL].get(controller)!;
-  controllerRef.handlers ??= new Map();
-
-  if (!controllerRef.handlers.has(fn)) {
-    controllerRef.handlers.set(fn, {});
-  }
-
-  const handlerRef = controllerRef.handlers.get(fn)!;
   handlerRef.passResponse = {
     paramName,
     index: parameterIndex,
@@ -33,9 +23,9 @@ export const Response: ParameterDecorator = (
 
   if (process.env.AURALIS_DEBUG) {
     console.debug("[Response]:", {
+      args: [target, propertyKey, parameterIndex],
       controller,
-      propertyKey,
-      parameterIndex,
+      fn,
       paramName,
     });
   }
